pyproject.toml
[tool.poetry]
name = "enecad-ia"
version = "0.1.0"
description = ""
authors = ["ds-digital-23 <DSWeb@dronessolution.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = "2.2.0"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
pydantic-settings = "^2.3.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "^4.1.3"
ultralytics = "^8.2.28"
aiohttp = "^3.9.5"
python-decouple = "^3.8"
gdown = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
